move LinkedContainerManager interface to better home
