+++ src/main/java/com/o19s/grandcentral/GrandCentralApplication2.java
@@ -10,7 +10,6 @@ import javax.servlet.DispatcherType;
 
 import com.o19s.grandcentral.dockercloud.DockercloudRegistry;
 import com.o19s.grandcentral.dockercloud.StackManager;
-import com.o19s.grandcentral.kubernetes.LinkedContainerManager;
 import com.o19s.grandcentral.servlets.PodProxyServlet;
 import com.o19s.grandcentral.servlets.PodServletFilter;
 //import com.o19s.grandcentral.healthchecks.ContainerRegistryHealthCheck;
