+++ src/main/java/com/o19s/grandcentral/dockercloud/StackManager.java
@@ -35,9 +35,9 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
+import com.o19s.grandcentral.LinkedContainerManager;
 import com.o19s.grandcentral.http.HttpDelete; // IMPORTANT, allows DELETE requests with bodies
 import com.o19s.grandcentral.kubernetes.Pod;
-import com.o19s.grandcentral.kubernetes.LinkedContainerManager;
 
 /**
  * Manages all stacks present within a namespace
