+++ src/main/java/com/o19s/grandcentral/kubernetes/PodManager.java
@@ -46,6 +46,7 @@ import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
+import com.o19s.grandcentral.LinkedContainerManager;
 import com.o19s.grandcentral.http.HttpDelete; // IMPORTANT, allows DELETE requests with bodies
 
 /**
