+++ src/test/java/com/o19s/grandcentral/kubernetes/PodManagerTest.java
@@ -10,6 +10,7 @@ import org.junit.Rule;
 import org.junit.Test;
 
 import com.github.tomakehurst.wiremock.junit.WireMockRule;
+import com.o19s.grandcentral.LinkedContainerManager;
 
 ;
 /**
